+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-Develop
+stats branch
+ - Implemented stats submethod
+ - Added deprecationwarning to Signal.calc_dbfs and Signal.calc_dbspl
+
+Develop branch
  - Added the option to directly  apply filters when generating partly correlated noise
  - Added the audiotools.cmplx_corr function which calculates the
    complex-valued correlation coefficent